bugfix(server): fix ThinkingEffort flag merging priority
Fix inverted logic in resolveRuleFlagsWithScenario where rule-level
ThinkingEffort settings were being overwritten by scenario defaults.

The previous condition `flags.ThinkingEffort != typ.ThinkingEffortDefault`
triggered when a rule explicitly set ThinkingEffort (to "off", "low", etc.),
causing the scenario's value to overwrite the user's configuration.

Corrected to only inject scenario-level ThinkingEffort when the rule
hasn't set it explicitly:
- Rule explicit setting > Scenario default > Client choice

Also:
- Add comprehensive test coverage for ThinkingEffort flag merging
- Merge rule_flags_clean_header_test.go into rule_flags_test.go
